Implement `halt` op; fix multi-program loads; misc
Now programs with no heartbeat are compiled to halt.

Problems existed which weren't previously detected:
New loads *pFirstItem was incorrect as memLen wasn't yet set.
This caused new loads' items to be be written in the wrong place.
To fix the problem, new loads' info is set in one call to switchToProg.

Another minor issue was the expectation of only N programs,
rather than an expectation for a minimum and maximum - now fixed.